Remove non-unique bucket name
S3 bucket names must be globally unique, removed
arbitrary bucketName parameter to auto-generate
a unique bucket name for new stacks.

Verified able to successfully deploy the stack
to an AWS account via `cdk deploy` after the change.

Can update to specify the bucketName parameter
for S3 buckets in the stack if we specify a bucket
name that is not used by any other AWS accounts.